extract useSdkContext
