+++ enterprise/frontend/src/embedding-sdk/store/index.ts
@@ -4,13 +4,7 @@ import type {
   Store,
   ThunkDispatch,
 } from "@reduxjs/toolkit";
-import { useContext } from "react";
-import {
-  ReactReduxContext,
-  type TypedUseSelectorHook,
-  useSelector,
-  useDispatch,
-} from "react-redux";
+import { useDispatch } from "react-redux";
 
 import type { SdkStoreState } from "embedding-sdk/store/types";
 import { mainReducers } from "metabase/reducers-main";
@@ -32,24 +26,10 @@ export const store = getStore(sdkReducers, null, {
   },
 }) as unknown as Store<SdkStoreState, AnyAction>;
 
-// eslint-disable-next-line no-literal-metabase-strings -- this string only shows in the console.
-export const USE_OUTSIDE_OF_CONTEXT_MESSAGE = `The useMetabaseAuthStatus hook must be used within a component wrapped by the MetabaseProvider`;
-
-export const useSdkSelector: TypedUseSelectorHook<SdkStoreState> = (
-  selector,
-  options,
-) => {
-  const context = useContext(ReactReduxContext);
-
-  if (!context) {
-    throw new Error(USE_OUTSIDE_OF_CONTEXT_MESSAGE);
-  }
-
-  return useSelector(selector, options);
-};
-
 export const useSdkDispatch: () => ThunkDispatch<
   SdkStoreState,
   void,
   AnyAction
 > = useDispatch;
+
+export { useSdkSelector } from "./use-sdk-selector";
